Rename to fabric-message-api-v1
