--- fabric-chat-api-v1/src/testmod/resources/fabric.mod.json
@@ -1,16 +1,0 @@
-{
-  "schemaVersion": 1,
-  "id": "fabric-chat-api-v1-testmod",
-  "name": "Fabric Chat API (v1) Test Mod",
-  "version": "1.0.0",
-  "environment": "*",
-  "license": "Apache-2.0",
-  "depends": {
-    "fabric-chat-api-v1": "*"
-  },
-  "entrypoints": {
-    "main": [
-      "net.fabricmc.fabric.test.chat.ChatTest"
-    ]
-  }
-}
