+++ fabric-message-api-v1/src/main/resources/assets/fabric-message-api-v1/icon.png
