+++ gradle.properties
@@ -12,7 +12,6 @@ fabric-api-base-version=0.4.8
 fabric-api-lookup-api-v1-version=1.6.4
 fabric-biome-api-v1-version=9.0.9
 fabric-blockrenderlayer-v1-version=1.1.17
-fabric-chat-api-v1-version=1.0.0
 fabric-command-api-v1-version=1.2.1
 fabric-command-api-v2-version=1.0.1
 fabric-commands-v0-version=0.2.18
@@ -32,6 +31,7 @@ fabric-key-binding-api-v1-version=1.0.17
 fabric-keybindings-v0-version=0.2.15
 fabric-lifecycle-events-v1-version=2.0.7
 fabric-loot-tables-v1-version=1.0.15
+fabric-message-api-v1-version=1.0.0
 fabric-mining-level-api-v1-version=2.1.5
 fabric-models-v0-version=0.3.14
 fabric-networking-api-v1-version=1.0.25
