+++ settings.gradle
@@ -16,7 +16,6 @@ include 'fabric-api-base'
 include 'fabric-api-lookup-api-v1'
 include 'fabric-biome-api-v1'
 include 'fabric-blockrenderlayer-v1'
-include 'fabric-chat-api-v1'
 include 'fabric-command-api-v2'
 include 'fabric-content-registries-v0'
 include 'fabric-crash-report-info-v1'
@@ -31,6 +30,7 @@ include 'fabric-item-groups-v0'
 include 'fabric-key-binding-api-v1'
 include 'fabric-lifecycle-events-v1'
 include 'fabric-loot-tables-v1'
+include 'fabric-message-api-v1'
 include 'fabric-mining-level-api-v1'
 include 'fabric-models-v0'
 include 'fabric-networking-api-v1'
